+++ sample/src/main/java/com/eoscode/springapitools/sample/configuration/SpringApiToolsScanConfig.java
@@ -1,4 +1,4 @@
-package com.eoscode.springapitools.sample.config;
+package com.eoscode.springapitools.sample.configuration;
 
 import com.eoscode.springapitools.config.SpringApiToolsScan;
 import org.springframework.context.annotation.Configuration;
